fix(api): isolate docker runtime from host socket
